fix(framework): reply instead of dropping stale relocation and session frames
Four e2e scenarios (ST-D1, ST-B3, ST-B4, ST-E1A) failed because the runtime
refused or silently dropped work it should have accepted or answered.

Relocation to an explicitly addressed target applied the new-placement rules
that govern *selection*. A target weighted out of the candidate pool was
refused even though the caller had named it, so admission never happened.
CommitAsync and CompleteCommitAsync already passed
requireNewPlacementEligibility: false; the reservation and commit paths were
missing it. Graceful drain does not use placement weight, so this does not
weaken 28-graceful-drain-handoff.

A session frame naming a replaced incarnation was dropped inside a one-way
route handler: the exception reached the dispatch error sink and the caller
waited out its deadline. Spec 20 requires the relay to end with a typed stale
error, so the frame's own reply route now answers first. The kind is NotFound
(DoNotRetry) rather than Unavailable, because a replaced incarnation never
returns; 10-monitoring-errors forbids promoting generation-stale to its own
public kind, so no new error value was added.

Spec 15 said completion waits for durable cleanup while its own step list and
sequence diagram put completion first, and left the failing-leave case
unspecified. Both are settled: completion waits for the target joined callback
only, and the source leave notification is a one-way send that cannot block it.

Diagnostics are added at each point where a cause used to be discarded -
silent nulls in spot projection and target eligibility, swallowed exceptions in
deferred join and handoff commit, and the admission and seal round trips. They
cost one bool test when the switch is off: SpotDiscovery now takes an
interpolated string handler, so 68 call sites no longer build a string that is
thrown away, and the environment is read once instead of per call.

Full solution: 1831 tests, 0 failures.

+++ framework/languages/java/zlink-framework-binding-internal/src/main/java/systems/zlink/framework/runtime/internal/binding/spot/MeshNode.java
@@ -44,7 +44,7 @@ public interface MeshNode extends AutoCloseable {
     int routerHighWaterMark();
 
     /** Sets the routed admission queue high-water mark override. */
-    void setRouterHighWaterMark(int value);
+    void setRouterHighWaterMark(long value);
 
     /** Returns the per-owner mailbox message budget. */
     long mailboxMessageBudget();
